Adding missing tests
